pyproject.toml
[tool.poetry]
name = "nonebot-plugin-wwgachalogs"
version = "0.1.0"
description = ""
authors = ["irelia"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.3.1"
nonebot-adapter-onebot = "^2.4.3"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
